Fixed specular Liighting

+++ Camera.cpp
@@ -17,7 +17,7 @@ Camera::Camera(int width_, int height_, Vector3 eye_, Vector3 target_, Vector3 u
 	upper_left = -u + v;
 	horizontal = Vector3();
 }
-Ray Camera::getRay(int X, int Y)
+Ray Camera::getRay(double X, double Y)
 {
 	Ray retRay;
 	retRay.origin = origin;
@@ -26,8 +26,8 @@ Ray Camera::getRay(int X, int Y)
 	double udir = (2.0 * (double(X) / double(width)) - 1);
 	double vdir = 1.0 - 2.0*(double(Y) / double(height));
 	*/
-	double udir = double(X) / double(width);
-	double vdir = double(Y) / double(height);
+	double udir = (X) / double(width);
+	double vdir = (Y) / double(height);
 	direction += (udir * u * 2 - vdir * v * 2) ;
 	direction = normalize(direction);
 	retRay.direction = direction;
