+++ Camera.h
@@ -19,7 +19,7 @@ public:
 	double tanfov;
 	Camera();
 	Camera(int width_, int height_, Vector3 eye_, Vector3 target_, Vector3 up_, double fov_);
-	Ray getRay(int X, int Y);
+	Ray getRay(double X, double Y);
 };
 
 #endif
